Add basic support for git submodules
With this commit, you can add or update submodules in a commit that is
managed by guilt.  If you pop a commit that adds a submodule, the
entire directory will be removed.  If you push such a commit, the
submodule will be initialized and updated again.  When you push or pop
a patch that changes a submodule, the submodule will be updated.

The content of all submodules must be clean, or guilt will abort.
Since submodules may be removed by "guilt pop", guilt uses a stricter
check for cleanliness for submodules in order to avoid data loss: they
may not contain any untracked paths, unless they are ignored.

Since "git apply" doesn't apply changes to submodule, this commit also
adds a parser for patch files, currently implemented in a Python 3
script (extract-submodule.py).  It is unfortunate that guilt now
depends on Python.  Perhaps I'll re-implement this as a C program
instead at a future date; I already had Python code I could reuse for
this commit.  Consider this a proof-of-concept.  (The Python code
initially solved a more complex task, so it is overly complex for what
it currently does here.  But it works.)

Known issues:

 - Submodules named with unusual characters may break.
   Please use only safe ASCII letters.

 - Only add, refresh, push and pop have been properly tested.

The following changes were needed:

 - Teach pop_many_patches() to deinit submodules before popping a
   patch that adds that submodule.

 - Teach pop_many_patches() to run git submodule update on submodules
   that are changed by the patch series that is popped (unless there
   is also a patch that adds the submodule; in that case the deinit
   code is used instead).

 - Teach push_patch() to run "git submodule update --init" on any
   submodule that is created by a patch.

 - Teach push_patch() to apply changes to which commit of a submodule
   that should be checked out.

 - Teach must_commit_first() to check if any submodule is dirty.

The new test case t-072.sh tests the above functionality.  To
implement that test case, a few changes were needed to the test
scaffold code:

 - Teach cmd() and shouldfail() to use replace_path when echoing the
   command.

 - Teach replace_path to normalize the path name for a directory used
   by the t-072.sh tests.

Unfortunately, since .gitmodules contains the path to the upstream
module, t-072.sh needs to use a fixed path to the upstream repository.
This means only one user at a time can run the test.  If dotlockfile
is available, it will use that program to guarantee exclusive access.
Otherwise, you have to ensure that manually.

The extract-submodule.py file is based on commit 1ce8b8e59d04 of the
pdiffdiff repo.